Add a pattern datatype.
Core.hs and Rule.hs now handle a new pattern datatype, the aim is to
provide dot patterns by the end of the day. Patterns have two
constructors, PV for variables and constants (we discriminate them using
the environment of the rule) and PA for pattern applications.

For the sake of simplicity it is mandatory that dot patterns are given
at the beginning of pattern, as it is shown by the type of PA.

Currently, I am not quite sure of the interactions between dot patterns
and non-linear patterns, but Mathieu is planning to remove non-linear
patterns.