Abstract variables' type.
In decision tree, we were asking that the type of the constructors which
are matched on is the same as the type of representing variables. This
is silly, and made the module tricky to use. Now two different types can
be used for these objects a priori of different kinds.