Add weird constructors.
My idea is to remove the App code constructor in order to have lighter
pattern matching while defining code for rewrite rules. Because of the
restrictions made on rewrite rules, I think that the current
representation of applied constants will be sufficient to handle all
cases, this is still quite unclear.

Note that we make use of GADTs to describe what kind of applied constant
we are facing.

I still have a doubt about the Z constructor, it seems to be useful only
for pretty printing, will have to think about it.