Finalize new constructors handling.
Code generation is now aware of the new runtime and generates
appropriate code. I removed Var from the runtime and replaced it by a
smart constructor 'var' which generates a fresh new constructor name
based on the given integer. If we don't do this 'ap' will not be able to
handle applied variables while performing the conversion test.

All tests are still ok.