Generate code similar to the one in 'cons'.
I plan to use 6 different code kinds, first comes 'crule', which
represents a partially applied rule (i.e. a Lua function + its arity +
an array of arguments already provided), then comes 'ccon' which
represents an applied constant (i.e. a string and an array of
arguments), then we have 'cpi' for the dependent product, then 'clam'
for the usual abstraction, and 'ctype' and 'ckind'.

The code generated for rules is 'crule' except when the rewrite rule is
a definition (arity 0), in this case, the code aliased is directly
generated.