Partial fix of pattern matching compilation.
When pattern matching failed, an error was thrown. In fact we must
return an applied constant. This is used, for instance, when a rewrite
rule is applied to variables (for instance, to test the convertibility
of two terms).

The fix is partial since captured variables are still not handled
correctly.